moved things from header, cleaned up build files
