+++ src/aod/meson.build
@@ -1,8 +1,13 @@
 message('in aod')
 
+# TODO as this grows, I should add conditionals:
+# if s7 found, add these files
+# if imgui found, ... and so on
 aod_sources = files(
   's7.cpp',
   's7/c_primitives.cpp',
+  's7/imgui.cpp',
+  's7/repl.cpp',
 )
 
 aod_lib = library(
@@ -10,19 +15,15 @@ aod_lib = library(
   # sources: aod_sources,
   include_directories: include_directories('..'),
   sources: aod_sources,
+  links_width: imgui,
   dependencies : [
     dependency('s7', required: true),
+    dependency('imgui', required: true),
   ]
   )
 
 aod_dep = declare_dependency(
-   # one up, doing include "aod/s7.hpp" etc
   link_with: aod_lib,
+  # include from parent dir: to include "aod/s7.cpp" for example
   include_directories: include_directories('..'),
-  # pf.. i'm confused about the dependencies..
-  # library vs declare_dependency
-  dependencies : [
-    # if i skip this, the tests that include s7.h throw an error
-    dependency('s7', required: true),
-  ]
   )
